change(mayor):
Create playlist context.
Add like to track in Mylibrary context and create the reducer
Adapt frontend for the new enpoints
Fix the problems by handling images

+++ src/reducers/admin/genreAdmin/genreAdmin.actions.ts
@@ -11,7 +11,7 @@ export async function initGenresAction(dispatch: any) {
 				payload: response.genres
 			})
 		} else {
-			throw new Error()
+			message.warning(`Something went wrong`)
 		}
 	} catch (err) {
 		message.error('Server error')
@@ -21,17 +21,19 @@ export async function initGenresAction(dispatch: any) {
 export async function postGenreAction(dispatch: any, data: any, token: any, messageApi: any) {
 	try {
 		const postGenre: any = await api.postGenreApi({ name: data.name }, token)
-		const postGenreImage: any = await api.putGenreImageApi(postGenre.genre._id, data.image.file.originFileObj, token)
-		if (postGenreImage.status === 200) {
-			messageApi.destroy()
-			messageApi.success(`Genre '${data.name}' created`)
-			return dispatch({
-				type: GenreAdmin.POST_GENRE,
-				payload: postGenreImage.genre
-			})
-		} else {
-			messageApi.destroy()
-			messageApi.error('Server error')
+		if (postGenre.status === 200) {
+			const postGenreImage: any = await api.putGenreImageApi(postGenre.genre._id, data.image.file.originFileObj, token)
+			if (postGenreImage.status === 200) {
+				messageApi.destroy()
+				messageApi.success(`Genre '${data.name}' created`)
+				return dispatch({
+					type: GenreAdmin.POST_GENRE,
+					payload: postGenreImage.genre
+				})
+			} else {
+				messageApi.destroy()
+				message.warning(`Something went wrong`)
+			}
 		}
 	} catch (err) {
 		messageApi.destroy()
@@ -52,7 +54,7 @@ export async function deleteGenreAction(dispatch: any, genre: any, token: any, g
 			})
 		} else {
 			messageApi.destroy()
-			message.error('Server error')
+			message.warning(`Something went wrong`)
 		}
 	} catch (err) {
 		messageApi.destroy()
@@ -63,24 +65,17 @@ export async function deleteGenreAction(dispatch: any, genre: any, token: any, g
 export async function updateGenreAction(dispatch: any, data: any, genre: any, token: any, genresState: any, messageApi: any) {
 	try {
 		let newGenre;
-		let imagePublicId = null
 		if (data.image?.file.originFileObj) {
 			const genreImageToUpdate = await api.putGenreImageApi(genre._id, data.image.file.originFileObj, token)
 			newGenre = genreImageToUpdate
-			imagePublicId = data.imagePublicId
 		}
-		if (data.name !== genre.name) {
-			const genreToUpdate = await api.updateGenreApi(genre._id, { name: data.name, imagePublicId }, token)
+		if (data.name && data.name !== genre.name) {
+			const genreToUpdate = await api.updateGenreApi(genre._id, { name: data.name }, token)
 			newGenre = genreToUpdate
 		}
-		const findIndexGenre = genresState.genres.findIndex((item: any) => item._id === genre._id)
-		if (newGenre.genre) {
+		if (newGenre.status === 200 && newGenre.genre) {
+			const findIndexGenre = genresState.genres.findIndex((item: any) => item._id === genre._id)
 			genresState.genres[findIndexGenre] = newGenre.genre
-		} else {
-			if (data.name) genre.name = data.name
-			genresState.genres[findIndexGenre] = genre
-		}
-		if (newGenre.status === 200) {
 			messageApi.destroy()
 			message.success(`Genre updated`)
 			return dispatch({
@@ -89,10 +84,10 @@ export async function updateGenreAction(dispatch: any, data: any, genre: any, to
 			})
 		} else {
 			messageApi.destroy()
-			message.info(`Nothing to update`)
+			message.warning(`Something went wrong`)
 		}
 	} catch (err) {
 		messageApi.destroy()
-		message.info('Server error')
+		message.error('Server error')
 	}
 }
